Added a test case and updated some test cases
