+++ Lorem Ipsum XR Translator/Assets/Tests/Scripts.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e9bd06be107afd04e836f940133966eb
+guid: 396078a11a1b0304e8f905a08c4e928d
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
