+++ Lorem Ipsum XR Translator/Assets/Tests/Scripts/Util.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7ed9f5707619b9e42b41bb631f3fa77e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
