+++ Lorem Ipsum XR Translator/Assets/Tests/Scripts/Util/CacheTests.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: fe4f3157f6f8d074f9a3988eca3a0e14
-AssemblyDefinitionImporter:
+guid: 09fd508d9e874404084850b8f04bda08
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
