+++ Lorem Ipsum XR Translator/Assets/Tests/Scripts/Util/CacheTests/CacheTests.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 429d5e3e320d9e347abcece52b54ea0a
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
